+++ src/main/kotlin/com/github/xygeni/intellij/notifications/NotificationService.kt
@@ -1,5 +1,6 @@
 package com.github.xygeni.intellij.notifications
 
+import com.intellij.notification.NotificationAction
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.project.Project
@@ -7,8 +8,8 @@ import com.intellij.openapi.project.Project
 object NotificationService {
     private const val GROUP_ID = "Xygeni Notification Group"
 
-    fun notifyInfo(content: String, project: Project? = null) {
-        notify(content, NotificationType.INFORMATION, project)
+    fun notifyInfo(content: String, project: Project? = null, action: NotificationAction? = null) {
+        notify(content, NotificationType.INFORMATION, project, action)
     }
 
     fun notifyError(content: String, project: Project? = null) {
@@ -19,10 +20,18 @@ object NotificationService {
         notify(content, NotificationType.WARNING, project)
     }
 
-    private fun notify(content: String, type: NotificationType, project: Project?) {
-        NotificationGroupManager.getInstance()
+    private fun notify(
+        content: String,
+        type: NotificationType,
+        project: Project?,
+        action: NotificationAction? = null
+    ) {
+        val notification = NotificationGroupManager.getInstance()
             .getNotificationGroup(GROUP_ID)
             .createNotification(content, type)
-            .notify(project)
+        if (action != null) {
+            notification.addAction(action)
+        }
+        notification.notify(project)
     }
 }